MBS-10377: Label fields clearly on Set cover art page
This adds proper labels to the fields in the Set cover art tab.
It also adds the artist to the display (can't hurt to have) and
changes the dates/countries from being in a <li> (which gets
impractical if there's a lot of them) to just being comma-listed,
same as labels/catnos.