MBS-5479: Add an "Edit Relationships" tab
This adds an "Edit Relationships" tab to releases by the "Edit" tab,
in order to make the relationship editor easier to find for everyone.
Since most users are currently used to click the link on the sidebar
and it also makes sense in that context, this leaves it in there as well
 - we certainly can use as big a chance as possible for people
to add relationships anyway.