MBS-10345: Redirect away from /merge if not ready to merge (#1198)
Currently going to /merge with just one entity in the queue
loads a useless, semi-broken page. This changes it so that it just
doesn't work, either sending the user back (if gone through a button,
which can happen if the merge queue changed in another tab
in the meantime), or just sending them to the front page if they
typed /merge manually.