Convert the Release index page to React + add track paging
Fixes MBS-9921
Fixes MBS-8844
Fixes MBS-11703

In addition to the 10-medium limit we've already had, I've added a 100
track limit, which covers the case of absurdly large mediums. When a
medium like that is expanded, the track requests are paged (via a new
endpoint, /ws/js/tracks) and can be fetched 100 at a time. I provided a
no-script fallback as before.