MBS-10411: Force coordinates as floats in JSON WS
When serializing the JSON for places, we want latitude and longitude to
always be outputted as floats. Perl doesn't have explicit "string" or "number"
types; the conversion between them is done automatically depending on
context in which they're used. Thus adding 0 to these values changes the
context, and forces Perl to change its internal representation of the
values to numbers. The JSON serializer follows suit.