Use lodash in place of terable
This will make switching to exact object types by default in Flow
slightly easier (since the library requires some updates to support
that).

I originally added this library because it has the advantage that it
natively supports iterables, but we aren't likely to be using a lot of
those, at least not yet (and if we do, can probably get by with just
rolling our own generators/for..of loops where we need them, or add
the dependency on terable back if necessary).

Other than iterable support, there's not much point in keeping both
utility libraries around (and we certainly can't remove lodash).