MBS-8566: Reuse release AC when importing CD stubs (#1193)
If a CD stub is being imported that has track artists matching the
release artist already selected, it's very unlikely that the user
actually wants a different artist for the tracks. This assigns the
release AC to the tracks if the track artist name matches the
reduced artist credit for the release.