MBS-10355: Display track artist on recording pages (#1207)
This patch adds a "Track artist" column to recording pages. While
in some cases that will be the same as the release and recording artist,
IMO it makes sense to always show it for consistency rather than hiding
it only when every single track in the page happens to have the
same artist as its release or as the recording.

If the track artist and the recording artist differ, the class
'track-recording-clash' is added to the relevant td. This does nothing
on MBS itself at the moment, but it allows userscripts to make the clash
more visible if desired.