Move entity0, entity1 out of attributeValues
attributeValues is ultimately passed to expand2 to replace attribute
names in the link phrase. Since {entity0} and {entity1} appear in the
link phrase in the same way attributes do, it was convenient to just put
those in the same object to be picked up in the same way.

However, we need to start storing attributes as raw LinkAttrT objects
rather than rendered React elements, since the implementation for
MBS-7678 needs access to them. These will no longer have the same type
as `entity0` and `entity1` (which are typically rendered EntityLinks),
so storing them in the same object will cause a lot of unnecessary
friction with the type system, and make providing a consistent way to
access *only* attributes awkward.