Required attributes on orderable link types
This fixes the display of required attributes (those which have a min >
0 in the link_type_attribute_type table) on orderable link types (those
which exist in the orderable_link_type table).

No such combination currently exists, so this fixes an issue you
wouldn't ever find on the production site. I've added a test to
demonstrate that the code works.

Previously we'd include the required attribute in the link phrase, but
we'd use the name of the *root* link attribute type rather than the one
used on the relationship. We'd also still include the attribute in the
extra attributes display. Both of these issues are fixed.

The Flow type changes in linkedEntities.js (removing the `+` sigils)
were made because `mergeLinkedEntities` actually writes to these
objects, i.e. they aren't read-only.