Set source attribute in load_entities, if unset
With 42ad9a5, we have to set the source attribute explicitly now.
`load_entities` was modified to set the target, but not the source.

Fixes MBS-10418 and possibly other cases where we have a relationship
fetched via `get_by_id` which then has `load_entities` invoked on it.
(This is likely not many other places, if any. The vast majority of
places we load relationships use the form
`$c->model('Relationship')->load($source)` for obvious reasons, which
does set the source attribute on the loaded relationships correctly.)