MBS-10420: Remove /all from /edits/all link
We have nothing that deals with /edits/all - this only works
because the default for /edits/whatever is to ignore the whatever
and load /edits (except for /edits/open).

This seems to be because the action for the TT was edits/all, but
that's unrelated to the actual link that needs displaying.