Hide duplicate aliases from release recordings
Because of the issue in MBS-10428 alias for recording artists
weren't being shown at all. This seems like a great chance to
ensure the size of the returned data is smaller by doing the same
for recording aliases as we do for recording tags and genres:
hide them if the same artist is already on the recording or track AC.

Unlike for tags and genres, I'm not removing aliases from tracks
since these are currently shown and removing them might break
current usage. We should probably ensure they're not returned
for ws/3 though.