MBS-10433: Block removing areas in country_area
Any area in the country_area table can't be removed anyway,
since there are constraints at the database level blocking that.
We shouldn't allow entering an edit trying to remove a country area.

This makes the release country check in can_delete redundant, so
replacing that with a call to is_release_country_area.