MBS-10431: Remove 'yourmusic' from 7digital URLs
Apparently "A 7digital URL with 'yourmusic' is for a signed-in user
to download their purchases". Since that's clearly not what we
want to link to, we should just remove the 'yourmusic/' bit and the
repeated id at the end.

Since 7digital redirects to https anyway, we can do the same here.