MBS-10437: Block removing non-empty area
If an area has any relationships, it's safer not to allow removing it.
This also blocks removing it if it has open edits. It makes sure a
removal edit is not opened if the area is already being merged,
and since location editors are pretty much always going to be
autoeditors, they shouldn't have a problem approving most
other edits if needed.

Because all the previous checks for can_delete are subsets of this new
query, we can just remove them and only leave this check plus the
check for whether the area is a release country (where we still want to
block removing it even if it is otherwise empty!). Since the release
country check will apply to all areas used on releases, there is no need
to check the release_country table in this query.