JavaScript test coverage for Selenium
Generates JavaScript test coverage for our Selenium tests using
Istanbul. If all goes well, should be saved as a build artifact on
CircleCI for viewing.

The way we handle browser alerts in our Selenium tests has changed.
The relevant alert for us is the one in the release editor that prevents
you from navigating away and losing your changes. When this appears, the
webdriver doesn't allow for the execution of any scripts until it's
handled. (We have to execute a script to retrieve the errors object from
the page, and now the coverage object.) We previously detected this by
attempting to execute our script, and checking for an
UnexpectedAlertOpenError exception (in which case we accepted the alert
and retried the script). Now that we have to attempt to retrieve the
coverage object too, this is a pain to perform in both places. A better
solution, which I've implemented here, is to require each Selenium test
to explicitly accept or dismiss alerts that appear with a new
`handleAlert` command. Thus tests should now handle any alerts that are
expected, and unexpected alerts will fail the test. (But note that to
trigger the release editor alert, we have to navigate away from the page
first, so this must be the second-to-last step of those tests.)