MBS-10440: Remove series/delete
series/MBID/delete works, and enters edits, but is unlinked.
This seems unintentional, given that no checking is made
at all to whether the series is indeed empty enough to delete
(sub can_delete { 1 }) and yet the sub delete { } block specifically
says in a comment "No deleting relationship-related stuff because it
should probably fail if it's trying to do that".
This gets rid of that option,
and just let the cleanup tool do its job.