MBS-10456: Show more info when entering artist merges
This shows the artist type, gender and area in artist merge pages.
Since this made the ArtistMerge and RecordingMerge pages very similar,
I refactored them a bit to share some more code.
Eventually we might want to have a general EntityMerge page
with options instead, but this is good enough for now IMO.