MBS-9654: Show expanded ACs on recording merges
When merging recordings, it's pretty useful to be able to see
whether, say, the same artist credit name hides a different artist (and
the disambiguation comment for the artist).
This makes both the list during the merge process and the list on the
edit page itself display show expanded artist credits in the artist
cell to make these things more visible than just on hover.