MBS-10089: Limit Works tab to written/performed works
In general, an artist's work tab is meant to show the works the artist
has performed or written. It doesn't make a lot of sense to show works
linked with relationships such as "named after" or "commissioned"
in this tab.
This limits the works listed in the tab to those with relationships in
@WRITER_RELATIONSHIP_GIDS. All other works will still be shown
in the artist's relationships page.

Fixing the test to use one of the writer relationships.