MBS-4299: Show expanded AC in merge release warning (#1162)
When merging releases and recordings, if the recording artists
differ, the user is shown the recordings with different artists,
so that they can make sure the merge seems correct. This does
not show disambiguations, which can lead to a confusing situation where
a user is told "these recordings have different artists" but the
only info shown is "Foo by Bar" for both recordings.

This changes the display to use the list we already use for edit medium
edits, where every artist in an AC is shown separately under the line
if they have a disambiguation.