Remove OauthTypeT
This allows us to remove a use of `any`.

While we could also fix this by keeping OauthTypeT but checking that
it's equal to 'installed' or 'web' in handleOauthTypeChange, that sort
of check would be dead weight, and keeping OauthTypeT doesn't actually
provide any type safety here: by casting the string from the select to
`any` and then to OauthTypeT, we're trusting what we get from the
browser in any case.