Remove `any` from entityHref
The main "fix" here is that LinkableEntity must be a disjoint union
distinguished by the entityType property.

The changes to MinimalCoreEntityT and MinimalAnnotatedEntityT (to make
them exact) were part of an alternate fix I was working on, and not
strictly necessary anymore, but I kept them because they're technically
correct.