Delete reactTextContent
This was used by EntityLink to determine whether there was a name
variation when `content` was a React element (rather than a string). It
tried to recurse into the React element tree and figure out the actual
text content of the element, sort of like `.textContent` would do in a
web browser. But this didn't work at all if the React element was a
custom component: by definition it's not rendered yet, so it's
impossible to know what the text content could be.

I also found only one actual case where this might be needed, and
simply changed that to set nameVariation directly. We still set
nameVariation automatically in cases where `content` is a string.

Removes a usage of the `any` Flow type.