Add Flow types to more Jed/gettext modules
Fixes a use of `any` in wrapGettext.js.

Note: flow-typed/npm/jed_v1.x.x.js is our own definition, not from
the flow-typed repository, so local modifications are fine. We use a
fork of Jed (on github at mwiencek/Jed), so perhaps it's better to move
the types there as a Flow declarations file.