MBS-4091: Avoid breaking lists while trimming (#1160)
From annotation formatting:
Lists: tab or 4 spaces and:
* bullets or 1., a., A., i., I. numbered items (rendered with 1.)

We are trimming all whitespace at the start and end of the
annotation though (FormHandler's default), meaning if the
annotation starts with a list the formatting for the first
element will be broken. This overrides the default trim method with one
that only trims the ending, to avoid this situation.