MBS-4776: Require confirmation for VA track artist (#1214)
We have a warning for Various Artists used as a track artist, but people
keep wrongly selecting it. This adds a checkbox so that they have to
explicitly confirm they want to use VA, which hopefully will make them
think twice about it (we can't disallow it completely since
it *is* correct in some rare cases).