MBS-7379: Add median duration to works
It's pretty useful to be able to see at a glance what's
the expected duration for recordings of a specific work.

This ignores recordings that are marked as partial recordings,
as part of a medley, as covers or are related
to more than one work.

Since the data used to calculate this can be dodgy (a lot of
partial recordings are missing that attribute, for example)
this displays an InformationIcon letting people know it can
be untrustworthy and indicating the most likely fix.