Fix ENTITY_CACHE_TTL 0
It's documented in lib/DBDefs.pm.sample that "[a] value of 0 [for
ENTITY_CACHE_TTL] indicates that no expiration is set." The
implementation of this was broken, however, because setting the EX
option on the key to 0 causes Redis to throw an error ("invalid
expire time"), so the key is never set.

This code predates our use of Redis for caching, so likely broke when we
replace memcached with it in 32bc1f094e5159d1609676e7bfa122af3ea19c95.