MBS-10524: /account/change-password is broken unless logged in (#1327)
The issue is ultimately that UserAccountTabs requires a user to
function. It doesn't make sense to use user/profile/layout.tt at all
without a user, but it "worked" until the React conversion because TT
allows magic behavior.

This patch uses the normal layout if the user isn't logged in.