MBS-10322: Port 'fab pot' to shell script
Replace 'pot' Fabric task with new 'po/update_pot.sh' script and
'--commit' option. It is a derivative of 'po/update_translations.sh'
and has the same behavior:

- Prevent incidentally committing pre-existing local changes
- No more ask to confirm commit message before committing
- Show commit statistics if any, print a notice otherwise
- Test (remote) database connection before querying it

Plus, by Michael: Set MB_POT_DB=PROD_STANDBY in update_pot.sh

The intent of this script is to update pot files using the production
database, so the developer is now required to have a PROD_STANDBY
database configured in lib/DBDefs.pm which does that, via SSH tunnel or
otherwise.