MBS-10522: Save subscriptions to merge destination
This was supposed to happen already, but merge_entities was running
after the around do_merge code deleted the subscriptions, so there was
nothing left for merge_entities to move. This moves it to run as part of
the around do_merge section of MergeSubscription, which seems sensible
given it's as much part of the subscription merge as the deletion is.