Rename merge_entities to transfer_to_merge_target
The method currently named merge_entities under Data::Subscription
doesn't merge any entities. All it does is transfer the subscriptions
from the entities being merged away to the entity that is the target
of the merge. As such, transfer_to_merge_target seems like a much
clearer, self-explanatory name for the method.