MBS-10562: Add phrases for future (sidebar) dates
This only adds the option for cases that seem likely to happen and
where we have custom phrases already. The opening and closing of a
place can be announced ahead of time, and so can the dissolution
of a group (but it would be strange to announce the foundation
of a group before it happens, very dodgy to announce the death of
a person and pretty much impossible to specify the birth of one).