MBS-10553: Translate report reason only if needed (#1346)
We currently always translate the report reasons whenever we use them.
This makes a lot of sense when we're doing it to show them to the user
(who will want them on their language), but not so much when sending
them with the report email (which goes to the MeB team who might not
speak said language). As such, this moves the moment the reasons are
translated from the %REASONs hash (which is used in both cases) to the
specific options_reason map for the form creation.