MBS-10568: Show number of discIDs on discID tab
Same as we do for cover art already. It is useful to know whether
a release already has discIDs or not without having to click and
visit its discIDs tab. This shows the number if the release can have
discIDs, but skips adding anything new if the release can't have
discIDs at all.