MBS-10567: Limit series types as per DB constraint (#1351)
When creating a new series type, you can currently select any
entity type for the items that is allowed for collections. The list of
types allowed for series is different though, meaning that the system
will allow you to try to create an artist series type, then fail at the
database level when it hits a constraint. This makes it so that only
entity types allowed by the allowed_series_entity_type constraint
are shown as options for new series types, mirroring the way it's done
for collections.