MBS-10571: Fix "localized" ModBot email display (#1358)
ModBot's notes can be localized since d2f9ced using a special syntax.
Where it was used, the notes weren't being mailed correctly, instead
exposing the internal syntax to the recipient.

This passes the ModBot note text through `Entity::EditNote::localize`
before sending out the email. `editor_id` and `text` are the only
attributes needed to run that method.

The message is sent in English (via a new `run_without_translations`
method in `Server::Translation`) because we don't know the language of
the recipient. This is the behavior it's always had, of course.