Fix MBS-10572: Have one Jed bundle per language (#1359)
Prior to this patch, we had separate Jed data bundles per language per
domain, and any page that used a domain other than mb_server had to
specify that so it could be loaded. This was horrible, because we'd
have to remember to update these domain lists on every change, and
before we did, had to figure out what pages the change affected.

For example, re MBS-10572 an l_countries call was introduced into the
ReleaseEvents component, which it turns out is used on a lot of pages
that don't have the countries domain loaded.

This patch combines all domain-specific Jed data bundles into a single
bundle per language. These new bundles are around 500KB before gzip and
loaded on every page if a language is selected. Some unnecessary data
will be loaded on some pages now, but on a lot of edit pages we had to
load most or all of the domains anyway; this at least reduces the number
of HTTP requests in those cases. In other cases, we don't have to worry
about loading them now.

There is a potential for future work in reducing the size of these.