Generate source maps in production (#1336)
This generates source maps (mapped to the original source) in
production, to be hosted on StaticBrainz. Sentry can use these to
display better stack traces.

For development, I checked the difference in build time between
source-map and cheap-source-map, and it was negligible (less than a
second or two for me). Didn't check for rebuilds, though.