MBS-10576: Support errors with HTML
This allows passing forms errors containing HTML. We interpret
the errors as HTML only when specifically requested to ensure
we still scape the appropriate characters otherwise.

For now, this changes only form_row_select since that's the one
that already has one error using HTML, but we might want to expand
it in the future (if we add links to documentation in error texts,
for example).

Even though this isn't yet needed anywhere that uses the React
components, I'm adding it there already for parity - we'll need it
once we convert release/merge anyway.