MBS-10579: Unbreak es-es/el-gr translations
The code and associated comment here don't make sense to me. The
language used in the Jed files should match what's in DBDefs, since
that's what gets set in the lang cookie and is what's used to fetch the
file.

I'd also be in favor of changing these to es/el (without country codes)
in MB_LANGUAGES in production and removing the symlinks - they were to
work around a bug in Locale::Util that's been fixed for years. But in
that case we'd still remove this code as unused.