MBS-10584: Collapse work artists if too many
For some works (mostly classical) there can be dozens or hundreds
of artists listed, making proper navigation of any work list including
them pretty much impossible. This makes it so that only the top four
artists are shown, with an option to see all, using the same general
idea as MBS-10424.