Support running Selenium tests in Firefox
Chrome doesn't allow proxying localhost requests (and I couldn't get
`proxy-bypass-list="<-loopback>"` to work at all on macOS), and that's
needed for the Selenium tests to work if plackup is listening on
localhost. You can create a custom localhost alias to work around it,
but Firefox also works as a substitute. Having more options is good in
any case, since some people may not have (or want to install) Chrome in
the first place.