Install perl modules in Dockerfile.tests
This allows us to more easily use the image outside of CircleCI, where
we won't have to copy the cpanm steps to anymore (e.g. to Jenkins).
They'll already be installed in the image. It probably also speeds up
the CircleCI run, even though the modules were already cached there.

The downside is that we'll have to rebuild the image whenever we change
our cpanfile.snapshot. But that doesn't happen often.