MBS-10058: Convert merge_artists to React
There's no reason for this not to use ArtistList (which shows more info
about the artists and puts them in a table more consistent with the way
we display other merge edits). So I changed it to use that.

I added begin and end dates since they seem relevant for figuring
out if the artists are the same too.