MBS-10622: Highlight grouped link phrases with pending edits
If multiple phrases are grouped together, e.g. "mixer and recording
engineer," the target entity was only highlighted if the first
relationship in the group happened to have pending edits. We now merge
the pending edits flag of all target entity groups when phrase groups
are merged.

Further, we now highlight link phrases that correspond to relationships
with pending edits, in order to restore some information that was lost
with MBS-7678.