Run Selenium tests on Jenkins
There's too much going on in the Selenium test container right now for
the puny CircleCI VMs to handle. This should avoid the frequent page
crashes we're seeing in Chrome recently, which I believe are due to a
shortage in memory ever since all of the search services were added.