Recurse submodules at $MB_SOLR_TAG
We previously cloned the mb-solr repo with only --recursive. That cloned
the master branch, and as expected, checked out submodules as they were
set in master. In a separate command, we then moved to checkout the
desired $MB_SOLR_TAG, but that did *not* update the submodules to match
what was set for the tag.

To fix this, I've modified the 'clone' command to use --branch with
$MB_SOLR_TAG, so that the submodules are initialized to their correct
commits.