Merge CircleCI jobs
With Selenium tests moved to Jenkins, it doesn't make as much sense to
split the js/perl-and-pgtap jobs anymore. The js job only takes about a
minute to cycle through, less if you only count it running the actual
tests. We can run these serially in a single job while saving a bit of
time on container startup costs. This also allows for more containers to
be available for multiple PRs to build concurrently.