MBS-1921: Display edit link under annotations
This adds an extra "Edit annotation" link under annotations.
This link will make it clearer to users that they can edit the current
annotation, which isn't always that clear at the moment since
the only edit link is on the sidebar, very far away from the text.

Because the "View annotation history" link is already there sometimes,
I'm separating them with a slash when that's the case.
Seems cleaner than using brackets or periods.