MBS-10659: Always show entity credits in rel edits (#1398)
If one or both entities on the relationship have a credit, we want
to still display that credit in the edit relationship edit, even if the
edit itself doesn't change the credit (because it is still relevant
to know the relationship is under a specific credit). Currently,
the edit will display the main artist name at all times except
if the credits are changing.