Fix warning about artwork image suffix on homepage (#1394)
Even though `suffix` is not actually used on homepage,
not defining it was spamming website logs with this warning:

    Use of uninitialized value in concatenation (.) or string at lib/MusicBrainz/Server/Entity/Artwork.pm line 88

This patch sets `suffix` to `spoof` to silent this innocuous warning.

Note: The same warning can be triggered on edits history, but that is
for other reasons to be investigated. This patch doesn’t hide these.