MBS-10633: Allow changing username capitalization
It should be possible to change someone's username to the same
username, but with different capitalization. This changes the comparison
that checks if the username has changed to allow those
capitalization--only changes, by comparing normalized usernames instead.
The registration system doesn't allow two usernames to only differ
in capitalization in any case, so this should be perfectly safe.