MBS-10683: Combine moving map and focusing
An issue occurs sometimes with Leaflet that moves the marker
to the entered coordinates, but then zooms the map on (0, 0).
It seems to be a race condition of some kind between panTo and setZoom.
This moves from panTo + setZoom to setView (which allows passing
a zoom parameter, obviating the need for a second call to setZoom).
panTo actually uses setView internally anyway.