MBS-10677: Return entity types for WS entities in rels and ACs
For some entity types (works, places) we were returning their types
on non-toplevel situations, but weren't actually loading them for
relationships, so they misleadingly appeared as null on JSON.

There doesn't seem to be any reason to not show the types here -
it doesn't make the response much larger and it's very useful data.
As such, the best option is to actually always show the types
for all entity types. Same applies for artists in artist credits.