MBS-10695: Show language names on work lists
We currently show the code plus the main language on hover,
but the code is unclear to most users and we do have enough
space here to show the whole name.

This moves the code to a data-iso-639-3 attribute, in case people
want to be able to access it still.