MBS-10607: Add context to (unknown) strings (#1400)
We have plenty of "(unknown)" strings, and the thing that is unknown
can be one of many. Depending on the thing, though, the translation
should vary (for example, a release format is "desconocido" in Spanish
but a release itself is "desconocida" - gender variation). This adds
context to the different strings to make it possible to translate them
separately.