Skip exporting undefined or empty setlist to JSON
When exporting an Event to JSON, `setlist` may be not defined as event
search indexes do not store it, then `format_setlist` triggers warnings:

    Use of uninitialized value $text in substitution (s///) at lib/MusicBrainz/Server/Filters.pm

Even though `setlist` has default value NULL in the database schema,
it is actually stored as empty string because of Entity/Edit::Event.

    https://github.com/metabrainz/musicbrainz-server/blob/v-2020-02-04/admin/sql/CreateTables.sql#L694
    https://github.com/metabrainz/musicbrainz-server/blob/v-2020-02-04/lib/MusicBrainz/Server/Entity/Event.pm#L29
    https://github.com/metabrainz/musicbrainz-server/blob/v-2020-02-04/lib/MusicBrainz/Server/Edit/Event/Create.pm#L31

This patch skips serializing `setlist` if it is not defined (search) or
if it is an empty string (de facto default value in Perl and database).