Ignore case for sorting entities being merged
We generally would expect, say, "guitar" to sort before "Portuguese
guitar" when merging, but standard sort isn't case-insensitive.
This creates a sortByEntityName function that sorts using toLowerCase,
and changes the merge pages to use it.
In cases where the names are the same, we sort by id to at least
ensure the entities keep their positions after a reload.