MBS-10723: Separate className for cell and header
The lengthClass for recording merges should only apply to cells,
but currently className applies to both cells and headers.
This uses the existing headerProps and creates an equivalent
cellProps for cell attributes.

Now that there's a separation between cell and header classes,
this also changes the tables so that classes apply only to headers
or cells, accordingly, mimicking what they originally did on the
TT files (but with added consistency).