MBS-10736: Add Napster autoselect + sidebar
Deals with the (older?) subdomains www. and app. as well as the
no-subdomain links by redirecting them to the us. version, which
seems to be the main one. These redirect to the local version for
the user, or to a "not available in your area" page if there is no
local version.

Redirects (art./alb./tra.) links are blocked to avoid duplicate
links, but the custom error should give the user a clear idea
of what to do with them.

There's a separate subdomain vivomusica. which seems to be for a
"VivoMusica by Napster" site. This doesn't currently deal with it
at all since it's not 100% a Napster link. If needed, it can be
allowed in the future.