MBS-10741: Make "in use" pages consistent
This changes some inconsistencies between RelationshipTypeInUse
and RelationshipAttributeTypeInUse. These two are effectively
the same, so it makes sense they look the same too.

The differences being changed are:
1) The heading is standardised to <h1>
2) The heading and title are made the same.
3) The text is made equivalent. This will break translations, but the
breaking is acceptable because the old wording was nonsensical:
it mentions the attribute being "in use by other relationships", but
this is not reached through a relationship page, so there's no such
thing as "other" relationships.

To avoid breaking translations, I didn't currently change the headings
to be both sentence case nor both all-words-uppercased, since that
seemed minor.