MBS-10107: Add overview page for rel attributes
There's no good reason why we don't have these, and it would be
good if relationship-attribute/$gid wasn't an error.

If we have them, it makes sense to go to them and not the main list
after creating or adding an attribute, so I made that change too,
and linked to them from the attributes list.