Remove instance refs from Autocomplete2
It turns out that using Autocomplete2 in the new relationship editor is
difficult with Flow, because the Item type isn't parameterized. A first
step in making that easier to achieve is removing the whole "instance"
ref on every component. This was a bit of a hack to begin with: the idea
was to avoid unnecessary allocations on every render, but it made the
code confusing (and potentially buggy) because instance.state became a
reference to the state of previous render, not the current one (though
this didn't matter much in event handlers).

I'll certainly feel more confident that there aren't weird edge cases
lurking around once it's implemented more in line with how the official
React docs prescribe. That we can more easily improve the types this way
is a nice bonus.