MBS-10640: Fix donation check (#1454)
This was only saying you needed to donate if the MeB site
returned a nag status of 1: "known editor, donation time has passed".
It was not saying it when it returned -1 ("we don't know of this editor,
has never donated before"). This change makes it so that only 0
("no-nag period applies currently") is shown as "you don't need
to donate".