Convert form_row_name_with_guesscase to React
Also converts the guess case options dialog with it.

I've only used the new components in the recording edit form as an
initial test.

The state handling is a bit messy at the moment, since we don't have a
top-level RecordingForm component yet (which is where the state would
normally go). But eventually things should work themselves out.