MBS-10760: Convert Remove track edit to React
Changes from the TT:

Added the edit classes to be more consistent and improve
the display of the data.

Added a mock recording to the display data when we don't have it
to avoid DescriptiveLink dying.

Implemented reusable HistoricReleaseList component to provide
consistency for all the similar sections of historic edits.
This means some edits here and elsewhere will go from showing
an empty "Releases:" -> ""  to "Releases:" -> "[removed]"