MBS-10762: Convert historic Remove Release edit to React
Changes from the TT:

Adds the "remove-release" class, for consistency.

Doesn't show the "Artist:" label if artistCredit doesn't exist.

Uses HistoricReleaseList for consistency.
That means it uses DescriptiveLink to show the releases' ACs. This might
seem redundant given the Artist field below, but that shows the artist
at the time the edit was entered, and is also sometimes missing.