MBS-10765: Convert historic Remove Disc ID edit to React
Changes from the TT:

Added the edit classes to be more consistent and improve
the display of the data.

Uses HistoricReleaseList for consistency.

This also included converting CDTocLink.

I also added the cdtoc_id to the mock CDTOC (even if we don't
currently use it) since entities are always expected to have an ID.