Return heading in fallback for Relationships table
We generally show headings *plus* a fallback message when needed.
For example, for places, the Events tab shows "Events" plus a message
if empty, but the Performances tab currently shows just the message.
This makes it so that the heading is also returned if RelationshipsTable
has nothing else to print and a fallback message exists.