MBS-9086: Move most relationships away from area overview
Area overviews are overloaded with relationships. They list almost every
relationship for the area, meaning popular ones like New York or London
have tens of thousands of relationships being loaded on the overview and
often just time out.

This splits the biggest offender (recording rels) into their own tab,
and also moves release and work rels away for consistency and to
leave only relationships directly relevant to the area itself (area-area
and area-url) on the overview.

The other entity types left in the 'show' action currently have no
area-related relationship types, except for instrument
(which has one,"instrument origin", which is unlikely to get
unreasonable amounts of usage for any area in particular).
We should still keep them there so that if we do add any relationship
types for those combinations the relationships will be shown *somewhere*
until we decide whether they should be moved to separate tabs.

It's worth mentioning that this doesn't really solve the problem
of very large pages full of relationships timing out (MBS-1178),
but at least it should make it possible to always reach the main page
for an area - its Recordings tab might not be so easy to load...