MBS-8412: Standardize pagination limits
We have different pagination limits for different edit pages,
which leads to situations like artist/mbid/edits showing 100 edits
but the edit search reached when clicking "Refine this search"
(which a user would no doubt expect to be equivalent, until they
refine it) only showing the first 25 of them.

This standardises all of those lists to 100, which is our default
pagination limit anyway.