MBS-10772: Show top 5 genres and other tags on sidebar
Currently, if the 5 most voted tags for an entity are non-genre tags
the sidebar will make it look as if the entity has no genre tags at all,
since only the top 5 tags overall are displayed there. This changes it
so that both the top 5 genres and the top 5 non-genre tags are displayed
on the sidebar - it doesn't take that much space and is useful to make
sure genres don't drown tags nor vice versa.