MBS-10778: Add CSRF tokens to user/admin forms
A valid token is now required in the request parameters in order to
process these forms. The token is embedded into the HTML, but
cross-origin requests can't read these due to the same-origin policy in
web browsers.

If the same form is loaded twice (in two different tabs, say), the first
will fail to submit due to the token being invalid. However, the
submitted data is preserved, and we display a message telling them to
resubmit it.

These haven't been added to edit pages in order to not break any
existing userscripts that might make use of cross-origin data
submissions (though it's doable to enforce this in the future in places
where seeding the forms without actually submitting them is possible).