Add React context for SanitizedCatalystContext
(N.B. Confusingly, "context" has two different meanings here: the
Catalyst context, and React context:
https://reactjs.org/docs/context.html).

Client-side components that are hydrated from the server sometimes need
access to the Catalyst context (which is passed around via React
context, if that makes sense). Since the Catalyst context may contain
sensitive information about the request or user and is ultimately
embedded into the page, we conservatively create a "sanitized" Catalyst
context with only the essential information needed by client scripts.

Curiously, there's never been a React context provider for the sanitized
(or unsanitized) Catalyst context on the client. What we've done is wrap
any `hydrate` calls with `withCatalystContext`. That caused `hydrate` to
include a sanitized version of `$c` in the `data-props` attribute read
on the client. But this was the only way to access `$c` on the client;
any subcomponents needing it had to have it passed directly by a parent
(in other words, we couldn't use <CatalystContext.Consumer> in a client
side component).

To allow client components to consume the sanitized context in the same
way server components can consume the unsanitized context, I've added a
new React context called SanitizedCatalystContext. To ensure this can be
accessed in any location, `hydrate` now unconditionally includes the
sanitized Catalyst context data into a new `data-context` attribute
alongside `data-props` on the container node. While hydrating, it then
wraps the hydrated element in a <SanitizedCatalystContext.Provider>
using the parsed `data-context` value.