MBS-10793: Convert historic Move Release edit to React
Changes from the TT:

Actually attempt to link the artists if possible, rather than just
printing their names. That might be because a fair amount of them
have their IDs missing (one example had Metallica with an ID of 0)
but we should still try to link them where we can.

Additionally, this actually tries to pass move_tracks to the template -
it was trying to load it from display_data but AFAICT it wasn't
being added to display_data at all. Not sure if it's ever not false,
but just in case.